requirements.txt
PyYAML
bs4
lxml
python-dotenv
google-generativeai
